Abuse Turbo
When "build_complete" action is received in before-stream-render grab the rendered turbo
frame and add it to the dom. Also need to invoke all the JS rendering code so that
had to be cleaned up and broken out into re-usable calls.

I could not get the turbo replace stuff to work as I wanted. This also has the advantage
of being able to trigger CSS transitions as I update the DOM.